schedule: let merge not easy to timeout (#1495)
* fix merge easy to timeout
